feat: `--file/.file` can load dirs
